requirements.txt
robotframework>=6.1.0
uiautomator2>=3.0.0
dashscope>=1.14.0
click>=8.1.0
PyYAML>=6.0
Pillow>=10.0.0
colorama>=0.4.6
requests>=2.31.0
